commit 2: UI and model setup
Developed some models for the Android project to use. Login is a
Singleton as to avoid needed to pass parameters through.

Diary is composed of some class variables and a collection of
Appointments- as you can’t have a “Diary”(American,rather than English,
usage of the term).

Appointments themselves consist simply of a title, date start, end and
unique id.